Move Client and Controller into own folders
